requirements.txt
antlr4-python3-runtime==4.8
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
bokeh==2.4.3
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cryptography==37.0.1
data-lunch-cli==1.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
greenlet==1.1.1
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.2.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter_core==4.11.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
notebook==6.4.12
numexpr==2.8.3
numpy==1.23.1
omegaconf==2.1.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
panel==0.14.0
param==1.12.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.20
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyct==0.4.8
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyviz-comms==2.0.2
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
Send2Trash==1.8.0
setuptools==63.4.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.39
stack-data==0.2.0
terminado==0.13.1
testpath==0.6.0
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.5.2
zipp==3.8.0
